models moved back to db folder, seed.py added
